Re-structure API to nef library
minor changes
