+++ project/UI/CLIKit/Support Files/CLIKit.h
@@ -1,0 +1,15 @@
+//  Copyright © 2019 The nef Authors.
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for CLI.
+FOUNDATION_EXPORT double CLIVersionNumber;
+
+//! Project version string for CLI.
+FOUNDATION_EXPORT const unsigned char CLIVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <CLI/PublicHeader.h>
+#import "NefMarkdown.h"
+#import "NefJekyll.h"
+#import "NefCarbon.h"
+
