--- project/UI/Common/Support Files/NefCommon.h
@@ -1,14 +1,0 @@
-//  Copyright © 2019 The nef Authors.
-
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for Common.
-FOUNDATION_EXPORT double CommonVersionNumber;
-
-//! Project version string for Common.
-FOUNDATION_EXPORT const unsigned char CommonVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <Common/PublicHeader.h>
-#import "NefMarkdown.h"
-#import "NefJekyll.h"
-#import "NefCarbon.h"
