--- project/UI/Common/Support Files/module.modulemap
@@ -1,4 +1,0 @@
-framework module Common {
-    umbrella header "NefCommon.h"
-    export *
-}
