+++ project/UI/JekyllPage/main.swift
@@ -1,7 +1,7 @@
 //  Copyright © 2019 The nef Authors.
 
 import Foundation
-import Common
+import CLIKit
 
 let scriptName = "nef-jekyll-page"
 let console = JekyllConsole()
